tech: refactor main.ts to import sync and async ipchandlers
