+++ src/backend/main.ts
@@ -12,8 +12,6 @@ import { join } from 'path'
 import {
   app,
   BrowserWindow,
-  clipboard,
-  dialog,
   globalShortcut,
   ipcMain,
   Menu,
@@ -24,70 +22,42 @@ import {
 } from 'electron'
 import { uuid } from 'short-uuid'
 import { autoUpdater } from 'electron-updater'
-import { cpus, platform } from 'os'
+import { platform } from 'os'
 import {
   appendFileSync,
   existsSync,
-  readdirSync,
-  readFileSync,
   rmSync,
-  unlinkSync,
   watch,
   writeFileSync
 } from 'graceful-fs'
 import Backend from 'i18next-fs-backend'
 import i18next from 'i18next'
-import { DXVK, SteamWindows, Winetricks } from './tools'
 import { GameConfig } from './game_config'
 import { GlobalConfig } from './config'
 import { LegendaryUser } from 'backend/storeManagers/legendary/user'
 import { GOGUser } from './storeManagers/gog/user'
-import setup from './storeManagers/gog/setup'
 import {
-  clearCache,
-  execAsync,
   getPlatformName,
   getStoreName,
   isEpicServiceOffline,
   handleExit,
-  openUrlOrFile,
-  resetApp,
-  showAboutWindow,
-  showItemInFolder,
   wait,
-  getShellPath,
-  writeConfig
+  checkGameUpdates
 } from './utils'
 import {
-  configPath,
   configStore,
   createNecessaryFolders,
-  customThemesWikiLink,
-  discordLink,
-  epicLoginUrl,
   eventsToCloseMetaMaskPopupOn,
   fixAsarPath,
-  fontsStore,
   gamesConfigPath,
-  githubURL,
-  hyperplaySite,
   icon,
   installed,
   isCLIFullscreen,
   isCLINoGui,
-  isFlatpak,
   isSteamDeckGameMode,
-  onboardLocalStore,
   publicDir,
-  setQaToken,
-  sidInfoUrl,
-  supportURL,
   tsStore,
-  twitterLink,
-  userHome,
-  weblateUrl,
-  wikiLink,
-  wineprefixFAQ
+  userHome
 } from './constants'
 import { handleOtp, handleProtocol } from './protocol'
 import {
@@ -99,8 +69,6 @@ import {
   logWarning
 } from './logger/logger'
 import { gameInfoStore } from 'backend/storeManagers/legendary/electronStores'
-import { getFonts } from 'font-list'
-import { runWineCommand, verifyWinePrefix } from './launcher'
 import shlex from 'shlex'
 import { initQueue } from './downloadmanager/downloadqueue'
 import {
@@ -110,35 +78,22 @@ import {
 } from './online_monitor'
 import { notify, showDialogBoxModalAuto } from './dialog/dialog'
 import { addRecentGame } from './recent_games/recent_games'
-import { callAbortController } from './utils/aborthandler/aborthandler'
-import { getDefaultSavePath } from './save_sync'
 import { initTrayIcon } from './tray_icon/tray_icon'
 import {
   createMainWindow,
   getMainWindow,
   sendFrontendMessage
 } from './main_window'
-import { addGameToLibrary } from './storeManagers/hyperplay/library'
 
-import * as HyperPlayGameManager from 'backend/storeManagers/hyperplay/games'
 import * as HyperPlayLibraryManager from 'backend/storeManagers/hyperplay/library'
-import * as GOGLibraryManager from 'backend/storeManagers/gog/library'
 import {
-  getGOGPlaytime,
   syncQueuedPlaytimeGOG,
   updateGOGPlaytime
 } from 'backend/storeManagers/gog/games'
 import { playtimeSyncQueue } from './storeManagers/gog/electronStores'
 import * as LegendaryLibraryManager from 'backend/storeManagers/legendary/library'
 import { getFlag, initLDClient } from './flags/flags'
-import {
-  autoUpdate,
-  gameManagerMap,
-  initStoreManagers,
-  libraryManagerMap,
-  sendGameUpdatesNotifications
-} from './storeManagers'
-import { legendarySetup } from 'backend/storeManagers/legendary/setup'
+import { gameManagerMap, initStoreManagers } from './storeManagers'
 
 import * as Sentry from '@sentry/electron'
 import { DEV_PORTAL_URL, devSentryDsn, prodSentryDsn } from 'common/constants'
@@ -148,10 +103,8 @@ import { initExtension } from './extension/importer'
 import { hpApi } from './utils/hyperplay_api'
 import {
   initializeCompatibilityLayer,
-  checkWineBeforeLaunch,
-  runWineCommandOnGame
+  checkWineBeforeLaunch
 } from './utils/compatibility_layers'
-import { isClientUpdating } from 'backend/updater/updater'
 
 /*
  * INSERT OTHER IPC HANDLERS HERE
@@ -170,22 +123,17 @@ import './metrics/ipc_handler'
 import 'backend/extension/provider'
 import 'backend/proxy/ipcHandlers'
 
-import './ipcHandlers'
-import './ipcHandlers/checkDiskSpace'
-
 import { metricsAreEnabled, trackEvent } from './metrics/metrics'
 import { hpLibraryStore } from './storeManagers/hyperplay/electronStore'
 import { libraryStore as sideloadLibraryStore } from 'backend/storeManagers/sideload/electronStores'
 import { backendEvents } from 'backend/backend_events'
 import { PROVIDERS } from 'common/types/proxy-types'
-import 'backend/ipcHandlers/quests'
-import 'backend/ipcHandlers/achievements'
+
+import './ipcHandlers/index'
+
 import 'backend/utils/auto_launch'
+
 import { hrtime } from 'process'
-import {
-  getEpicListingUrl,
-  getHyperPlayReleaseObject
-} from './storeManagers/hyperplay/utils'
 import { checkG7ConnectionStatus, postPlaySessionTime } from './utils/quests'
 
 import { gameIsEpicForwarderOnHyperPlay } from './utils/shouldOpenOverlay'
@@ -233,11 +181,6 @@ if (metricsAreEnabled()) {
 }
 
 import { logFileLocation as getLogFileLocation } from './storeManagers/storeManagerCommon/games'
-import { addNewApp } from './storeManagers/sideload/library'
-import {
-  getGameOverride,
-  getGameSdl
-} from 'backend/storeManagers/legendary/library'
 import getPartitionCookies from './utils/get_partition_cookies'
 
 import { formatSystemInfo, getSystemInfo } from './utils/systeminfo'
@@ -250,7 +193,6 @@ if (!app.isPackaged || process.env.DEBUG_HYPERPLAY === 'true') {
   })
 }
 
-const { showOpenDialog } = dialog
 const isWindows = platform() === 'win32'
 
 let mainWindow: BrowserWindow
@@ -258,17 +200,6 @@ let ignoreExitToTray = false
 
 ipcMain.on('ignoreExitToTray', () => {
   ignoreExitToTray = true
-})
-
-ipcMain.on('focusMainWindow', () => {
-  const mainWindow = getMainWindow()
-
-  if (!mainWindow) {
-    return
-  }
-
-  mainWindow.show()
-  mainWindow?.focus()
 })
 
 async function completeHyperPlayQuest() {
@@ -697,9 +628,6 @@ if (!gotTheLock) {
     return
   })
 }
-
-ipcMain.on('notify', (event, args) => notify(args))
-
 ipcMain.once('loadingScreenReady', () => {
   logInfo('Loading Screen Ready', LogPrefix.Backend)
 })
@@ -740,28 +668,6 @@ process.on('uncaughtException', async (err) => {
   }) */
 })
 
-let powerId: number | null
-
-ipcMain.on('lock', () => {
-  if (!existsSync(join(gamesConfigPath, 'lock'))) {
-    writeFileSync(join(gamesConfigPath, 'lock'), '')
-    if (!powerId) {
-      logInfo('Preventing machine to sleep', LogPrefix.Backend)
-      powerId = powerSaveBlocker.start('prevent-app-suspension')
-    }
-  }
-})
-
-ipcMain.on('unlock', () => {
-  if (existsSync(join(gamesConfigPath, 'lock'))) {
-    unlinkSync(join(gamesConfigPath, 'lock'))
-    if (powerId) {
-      logInfo('Stopping Power Saver Blocker', LogPrefix.Backend)
-      powerSaveBlocker.stop(powerId)
-    }
-  }
-})
-
 ipcMain.on('quit', async () => handleExit())
 
 // Quit when all windows are closed, except on macOS. There, it's common
@@ -784,235 +690,10 @@ app.on('open-url', (event, url) => {
   }
 })
 
-ipcMain.on('openExternalUrl', async (event, url) => openUrlOrFile(url))
-ipcMain.on('openFolder', async (event, folder) => openUrlOrFile(folder))
-ipcMain.on('openSupportPage', async () => openUrlOrFile(supportURL))
-ipcMain.on('openReleases', async () => openUrlOrFile(githubURL))
-ipcMain.on('openWeblate', async () => openUrlOrFile(weblateUrl))
-ipcMain.on('showAboutWindow', () => showAboutWindow())
-ipcMain.on('openLoginPage', async () => openUrlOrFile(epicLoginUrl))
-ipcMain.on('openDiscordLink', async () => openUrlOrFile(discordLink))
-ipcMain.on('openTwitterLink', async () => openUrlOrFile(twitterLink))
-ipcMain.on('openWinePrefixFAQ', async () => openUrlOrFile(wineprefixFAQ))
-ipcMain.on('openWebviewPage', async (event, url) => openUrlOrFile(url))
-ipcMain.on('openWikiLink', async () => openUrlOrFile(wikiLink))
-ipcMain.on('openSidInfoPage', async () => openUrlOrFile(sidInfoUrl))
-ipcMain.on('openCustomThemesWiki', async () =>
-  openUrlOrFile(customThemesWikiLink)
-)
-ipcMain.on('showConfigFileInFolder', async (event, appName) => {
-  if (appName === 'default') {
-    return openUrlOrFile(configPath)
-  }
-  return openUrlOrFile(path.join(gamesConfigPath, `${appName}.json`))
-})
-
-// Calls WineCFG or Winetricks. If is WineCFG, use the same binary as wine to launch it to dont update the prefix
-ipcMain.handle('callTool', async (event, { tool, exe, appName, runner }) => {
-  const gameSettings = await gameManagerMap[runner].getSettings(appName)
-  const { wineVersion, winePrefix } = gameSettings
-  await verifyWinePrefix(gameSettings)
-
-  switch (tool) {
-    case 'winetricks':
-      await verifyWinePrefix(gameSettings)
-      await Winetricks.run(wineVersion, winePrefix, event)
-      break
-    case 'winecfg':
-      runWineCommandOnGame(runner, appName, {
-        gameSettings,
-        commandParts: ['winecfg'],
-        wait: false
-      })
-      break
-    case 'runExe':
-      if (exe) {
-        const workingDir = path.parse(exe).dir
-        runWineCommandOnGame(runner, appName, {
-          gameSettings,
-          commandParts: [exe],
-          wait: false,
-          startFolder: workingDir
-        })
-      }
-      break
-  }
-})
-
-ipcMain.handle('runWineCommand', async (e, args) => runWineCommand(args))
-
 /// IPC handlers begin here.
 
-async function checkGameUpdates(runners: Runner[]): Promise<string[]> {
-  let oldGames: string[] = []
-  const { autoUpdateGames } = GlobalConfig.get().getSettings()
-  for (const runner of runners) {
-    let gamesToUpdate = await libraryManagerMap[runner].listUpdateableGames()
-    if (autoUpdateGames) {
-      gamesToUpdate = await autoUpdate(runner as Runner, gamesToUpdate)
-    }
-    oldGames = [...oldGames, ...gamesToUpdate]
-  }
-
-  sendGameUpdatesNotifications().catch((e) =>
-    logError(
-      `Something went wrong sending update notifications: ${e}`,
-      LogPrefix.Backend
-    )
-  )
-
-  return oldGames
-}
-
-ipcMain.handle(
-  'checkGameUpdates',
-  async (e, runners: Runner[]): Promise<string[]> => {
-    return checkGameUpdates(runners)
-  }
-)
-
-ipcMain.handle('requestSIWE', HyperPlayGameManager.requestSIWE)
-
-ipcMain.handle('getEpicGamesStatus', async () => isEpicServiceOffline())
-
-ipcMain.handle('getMaxCpus', () => cpus().length)
-
-ipcMain.handle('getAppVersion', () => app.getVersion())
-ipcMain.handle('isFullscreen', () => isSteamDeckGameMode || isCLIFullscreen)
-ipcMain.handle('isFlatpak', () => isFlatpak)
-ipcMain.handle('getGameOverride', async () => getGameOverride())
-ipcMain.handle('getGameSdl', async (event, appName) => getGameSdl(appName))
-
-ipcMain.handle('getPlatform', () => process.platform)
-
-ipcMain.handle('showUpdateSetting', () => !isFlatpak)
-
-ipcMain.on('clearCache', (event, showDialog, fromVersionChange = false) => {
-  clearCache(undefined, fromVersionChange)
-  sendFrontendMessage('refreshLibrary')
-
-  showDialogBoxModalAuto({
-    event,
-    title: i18next.t('box.cache-cleared.title', 'Cache Cleared'),
-    message: i18next.t(
-      'box.cache-cleared.message',
-      'HyperPlay Cache Was Cleared!'
-    ),
-    type: 'MESSAGE',
-    buttons: [{ text: i18next.t('box.ok', 'Ok') }]
-  })
-})
-
-ipcMain.on('resetApp', async () => {
-  resetApp()
-})
-
-ipcMain.on('resetExtension', async () => {
-  const extensionImporter = await import('@hyperplay/extension-importer')
-  extensionImporter.resetExtension(hpApi)
-  ipcMain.emit('ignoreExitToTray')
-  app.quit()
-})
-
-ipcMain.on('createNewWindow', (e, url) => {
-  new BrowserWindow({ height: 700, width: 1200 }).loadURL(url)
-})
-
-ipcMain.handle('isGameAvailable', async (e, args) => {
-  const { appName, runner } = args
-  return gameManagerMap[runner].isGameAvailable(appName)
-})
-
-ipcMain.handle('appIsInLibrary', async (event, appName, runner) => {
-  if (runner !== 'hyperplay') return false
-  return HyperPlayGameManager.appIsInLibrary(appName)
-})
-
-ipcMain.on('goToGamePage', async (event, gameId, action) => {
-  return sendFrontendMessage('goToGamePage', gameId, action)
-})
-
-ipcMain.on('navigate', async (event, appName) => {
-  return sendFrontendMessage('navigate', appName)
-})
-
-ipcMain.handle('getGameInfo', async (event, appName, runner) => {
-  // Fastpath since we sometimes have to request info for a GOG game as Legendary because we don't know it's a GOG game yet
-  if (runner === 'legendary' && !LegendaryLibraryManager.hasGame(appName)) {
-    return null
-  }
-  return gameManagerMap[runner].getGameInfo(appName)
-})
-
-ipcMain.handle('getExtraInfo', async (event, appName, runner) => {
-  // Fastpath since we sometimes have to request info for a GOG game as Legendary because we don't know it's a GOG game yet
-  if (runner === 'legendary' && !LegendaryLibraryManager.hasGame(appName)) {
-    return null
-  }
-  return gameManagerMap[runner].getExtraInfo(appName)
-})
-
-ipcMain.handle('getGameSettings', async (event, appName, runner) => {
-  try {
-    return await gameManagerMap[runner].getSettings(appName)
-  } catch (error) {
-    logError(error, LogPrefix.Backend)
-    return null
-  }
-})
-
-ipcMain.handle('getGOGLinuxInstallersLangs', async (event, appName) =>
-  GOGLibraryManager.getLinuxInstallersLanguages(appName)
-)
-
-ipcMain.handle(
-  'getInstallInfo',
-  async (event, appName, runner, installPlatform, channelNameToInstall) => {
-    try {
-      const info = await libraryManagerMap[runner].getInstallInfo(
-        appName,
-        installPlatform,
-        'en-US',
-        channelNameToInstall
-      )
-      if (info === undefined) return null
-      return info
-    } catch (error) {
-      logError(
-        error,
-        runner === 'legendary' ? LogPrefix.Legendary : LogPrefix.Gog
-      )
-      return null
-    }
-  }
-)
-
-ipcMain.handle('getUserInfo', async () => {
-  return LegendaryUser.getUserInfo()
-})
-
-// Checks if the user have logged in with Legendary already
-ipcMain.handle('isLoggedIn', LegendaryUser.isLoggedIn)
-
-ipcMain.handle('login', async (event, sid) => LegendaryUser.login(sid))
-ipcMain.handle('authGOG', async (event, code) => GOGUser.login(code))
-ipcMain.handle('logoutLegendary', LegendaryUser.logout)
-ipcMain.on('logoutGOG', GOGUser.logout)
 ipcMain.handle('getLocalPeloadPath', async () => {
   return fixAsarPath(join(publicDir, 'webviewPreload.js'))
-})
-
-ipcMain.handle('getAlternativeWine', async () =>
-  GlobalConfig.get().getAlternativeWine()
-)
-
-ipcMain.handle('readConfig', async (event, config_class) => {
-  if (config_class === 'library') {
-    await libraryManagerMap['legendary'].refresh()
-    return LegendaryLibraryManager.getListOfGames()
-  }
-  const userInfo = await LegendaryUser.getUserInfo()
-  return userInfo?.displayName ?? ''
 })
 
 ipcMain.handle('requestSettings', async (event, appName) => {
@@ -1065,42 +746,6 @@ ipcMain.handle('requestSettings', async (event, appName) => {
   return mapOtherSettings(config)
 })
 
-ipcMain.handle('toggleDXVK', async (event, { appName, action }) =>
-  GameConfig.get(appName)
-    .getSettings()
-    .then(async (gameSettings) =>
-      DXVK.installRemove(gameSettings, 'dxvk', action)
-    )
-)
-
-ipcMain.handle('toggleDXVKNVAPI', async (event, { appName, action }) =>
-  GameConfig.get(appName)
-    .getSettings()
-    .then(async (gameSettings) =>
-      DXVK.installRemove(gameSettings, 'dxvk-nvapi', action)
-    )
-)
-
-ipcMain.on('toggleVKD3D', (event, { appName, action }) => {
-  GameConfig.get(appName)
-    .getSettings()
-    .then((gameSettings) => {
-      DXVK.installRemove(gameSettings, 'vkd3d', action)
-    })
-})
-
-ipcMain.handle('writeConfig', (event, { appName, config }) =>
-  writeConfig(appName, config)
-)
-
-ipcMain.on('setSetting', (event, { appName, key, value }) => {
-  if (appName === 'default') {
-    GlobalConfig.get().setSetting(key, value)
-  } else {
-    GameConfig.get(appName).setSetting(key, value)
-  }
-})
-
 // Watch the installed games file and trigger a refresh on the installed games if something changes
 if (existsSync(installed)) {
   let watchTimeout: NodeJS.Timeout | undefined
@@ -1113,24 +758,6 @@ if (existsSync(installed)) {
     watchTimeout = setTimeout(LegendaryLibraryManager.refreshInstalled, 500)
   })
 }
-
-ipcMain.handle('refreshLibrary', async (e, library?) => {
-  if (library !== undefined && library !== 'all') {
-    await libraryManagerMap[library].refresh()
-  } else {
-    const allRefreshPromises = []
-    for (const runner_i in libraryManagerMap) {
-      allRefreshPromises.push(libraryManagerMap[runner_i as Runner].refresh())
-    }
-    await Promise.allSettled(allRefreshPromises)
-  }
-})
-
-ipcMain.on('logError', (e, err, options) =>
-  logError(err, { ...options, prefix: LogPrefix.Frontend })
-)
-
-ipcMain.on('logInfo', (e, info) => logInfo(info, LogPrefix.Frontend))
 
 let powerDisplayId: number | null
 let gamePlaySessionStartTimes: Record<string, bigint> = {}
@@ -1187,12 +814,6 @@ ipcMain.handle(
     await postPlaySession(appName, runner, sessionPlaytimeInMs)
   }
 )
-
-ipcMain.handle('isClientUpdating', async () => {
-  return isClientUpdating()
-})
-
-ipcMain.on('restartClient', () => autoUpdater.quitAndInstall())
 
 // get pid/tid on launch and inject
 ipcMain.handle(
@@ -1441,21 +1062,6 @@ ipcMain.handle(
   }
 )
 
-ipcMain.handle('openDialog', async (e, args) => {
-  const mainWindow = getMainWindow()
-  if (!mainWindow) {
-    return false
-  }
-
-  const { filePaths, canceled } = await showOpenDialog(mainWindow, args)
-  if (!canceled) {
-    return filePaths[0]
-  }
-  return false
-})
-
-ipcMain.on('showItemInFolder', async (e, item) => showItemInFolder(item))
-
 ipcMain.handle(
   'uninstall',
   async (event, appName, runner, shouldRemovePrefix, shouldRemoveSetting) => {
@@ -1550,10 +1156,6 @@ ipcMain.handle(
     })
   }
 )
-
-ipcMain.on('removeFromLibrary', (event, appName) => {
-  HyperPlayLibraryManager.removeFromLibrary(appName)
-})
 
 ipcMain.handle('repair', async (event, appName, runner) => {
   if (!isOnline()) {
@@ -1700,11 +1302,6 @@ ipcMain.handle(
   }
 )
 
-ipcMain.handle('kill', async (event, appName, runner) => {
-  callAbortController(appName)
-  return gameManagerMap[runner].stop(appName)
-})
-
 ipcMain.handle('updateGame', async (event, appName, runner): StatusPromise => {
   if (!isOnline()) {
     logWarning(
@@ -1752,71 +1349,6 @@ ipcMain.handle('updateGame', async (event, appName, runner): StatusPromise => {
   logInfo('finished updating', LogPrefix.Backend)
   return { status }
 })
-
-ipcMain.handle(
-  'changeInstallPath',
-  async (event, { appName, path, runner }) => {
-    await libraryManagerMap[runner].changeGameInstallPath(appName, path)
-    logInfo(
-      `Finished changing install path of ${appName} to ${path}.`,
-      LogPrefix.Backend
-    )
-  }
-)
-
-ipcMain.handle('egsSync', async (event, args) => {
-  return LegendaryLibraryManager.toggleGamesSync(args)
-})
-
-ipcMain.handle('syncGOGSaves', async (event, gogSaves, appName, arg) =>
-  gameManagerMap['gog'].syncSaves(appName, arg, '', gogSaves)
-)
-
-ipcMain.handle(
-  'syncSaves',
-  async (event, { arg = '', path, appName, runner }) => {
-    if (runner === 'legendary') {
-      const epicOffline = await isEpicServiceOffline()
-      if (epicOffline) {
-        logWarning(
-          'Epic is offline right now, cannot sync saves!',
-          LogPrefix.Backend
-        )
-        return 'Epic is offline right now, cannot sync saves!'
-      }
-    }
-    if (!isOnline()) {
-      logWarning('App is offline, cannot sync saves!', LogPrefix.Backend)
-      return 'App is offline, cannot sync saves!'
-    }
-
-    const output = await gameManagerMap[runner].syncSaves(appName, arg, path)
-    logInfo(output, LogPrefix.Backend)
-    return output
-  }
-)
-
-ipcMain.handle(
-  'getDefaultSavePath',
-  async (event, appName, runner, alreadyDefinedGogSaves) => {
-    return Promise.race([
-      getDefaultSavePath(appName, runner, alreadyDefinedGogSaves),
-      wait(15000).then(() => {
-        return runner === 'gog' ? [] : ''
-      })
-    ])
-  }
-)
-
-ipcMain.handle(
-  'checkHyperPlayAccessCode',
-  async (_e, licenseConfigId: number, accessCode: string) => {
-    return HyperPlayGameManager.validateAccessCode({
-      accessCode,
-      licenseConfigId
-    })
-  }
-)
 
 // Simulate keyboard and mouse actions as if the real input device is used
 ipcMain.handle('gamepadAction', async (event, args) => {
@@ -1922,103 +1454,6 @@ ipcMain.handle('gamepadAction', async (event, args) => {
   }
 })
 
-ipcMain.handle('getFonts', async (event, reload) => {
-  let cachedFonts = fontsStore.get('fonts', [])
-  if (cachedFonts.length === 0 || reload) {
-    cachedFonts = await getFonts()
-    cachedFonts = cachedFonts.sort((a, b) => a.localeCompare(b))
-    fontsStore.set('fonts', cachedFonts)
-  }
-  return cachedFonts
-})
-
-ipcMain.handle(
-  'runWineCommandForGame',
-  async (event, { appName, commandParts, runner }) => {
-    const gameSettings = await gameManagerMap[runner].getSettings(appName)
-
-    if (isWindows) {
-      return execAsync(commandParts.join(' '))
-    }
-    const { updated } = await verifyWinePrefix(gameSettings)
-
-    if (runner === 'gog' && updated) {
-      await setup(appName)
-    }
-    if (runner === 'legendary' && updated) {
-      await legendarySetup(appName)
-    }
-
-    // FIXME: Why are we using `runinprefix` here?
-    return runWineCommandOnGame(runner, appName, {
-      commandParts,
-      wait: false,
-      protonVerb: 'runinprefix'
-    })
-  }
-)
-
-ipcMain.handle('getShellPath', async (event, path) => getShellPath(path))
-
-ipcMain.handle('clipboardReadText', () => clipboard.readText())
-
-ipcMain.on('clipboardWriteText', (e, text) => clipboard.writeText(text))
-
-ipcMain.handle('getCustomThemes', async () => {
-  const { customThemesPath } = GlobalConfig.get().getSettings()
-
-  if (!existsSync(customThemesPath)) {
-    return []
-  }
-
-  return readdirSync(customThemesPath).filter((fileName) =>
-    fileName.endsWith('.css')
-  )
-})
-
-ipcMain.handle('getThemeCSS', async (event, theme) => {
-  const { customThemesPath = '' } = GlobalConfig.get().getSettings()
-
-  const cssPath = path.join(customThemesPath, theme)
-
-  if (!existsSync(cssPath)) {
-    return ''
-  }
-
-  return readFileSync(cssPath, 'utf-8')
-})
-
-ipcMain.on('addNewApp', (e, args) => addNewApp(args))
-
-ipcMain.handle('removeApp', async (e, args) => {
-  gameManagerMap[args.runner].uninstall(args)
-})
-
-ipcMain.handle('launchApp', async (e, appName, runner) =>
-  gameManagerMap[runner].launch(appName)
-)
-
-ipcMain.handle('installSteamWindows', async () => SteamWindows.installSteam())
-
-ipcMain.handle('isNative', (e, { appName, runner }) => {
-  return gameManagerMap[runner].isNative(appName)
-})
-
-ipcMain.handle('pathExists', async (e, path: string) => {
-  return existsSync(path)
-})
-
-ipcMain.handle(
-  'getPlaytimeFromRunner',
-  async (e, runner, appName): Promise<number | undefined> => {
-    if (runner === 'gog') {
-      return getGOGPlaytime(appName)
-    }
-
-    return
-  }
-)
-
 /*
   Other Keys that should go into translation files:
   t('box.error.generic.title')
@@ -2049,37 +1484,6 @@ backendEvents.on(
   }
 )
 
-ipcMain.on('openHyperplaySite', async () => openUrlOrFile(hyperplaySite))
-
-ipcMain.on('reloadApp', async () => {
-  for (const win of BrowserWindow.getAllWindows()) {
-    win.loadURL(win.webContents.getURL())
-  }
-})
-
-ipcMain.handle('addHyperplayGame', async (_e, projectId) => {
-  await addGameToLibrary(projectId)
-})
-
-ipcMain.handle('getEpicListingUrl', async (_e, projectId) =>
-  getEpicListingUrl(projectId)
-)
-
-ipcMain.handle(
-  'isGameHidden',
-  async (_e, gameId) =>
-    !!configStore
-      .get('games.hidden', [])
-      .find(({ appName }) => appName === gameId)
-)
-
-ipcMain.handle('unhideGame', async (_e, gameId) => {
-  const hiddenGames = configStore.get('games.hidden', [])
-  const newHiddenGames = hiddenGames.filter(({ appName }) => appName !== gameId)
-  configStore.set('games.hidden', newHiddenGames)
-  sendFrontendMessage('refreshLibrary', 'hyperplay')
-})
-
 function watchLibraryChanges() {
   // workaround for https://github.com/sindresorhus/electron-store/issues/165
   sideloadLibraryStore.onDidChange('games', (newValue) =>
@@ -2092,34 +1496,6 @@ function watchLibraryChanges() {
   })
 }
 
-ipcMain.on('openGameInEpicStore', async (_e, url) => {
-  if (url.startsWith('https://store.epicgames.com/'))
-    sendFrontendMessage('navToEpicAndOpenGame', url)
-})
-
-ipcMain.on('setQaToken', (_e, qaToken) => {
-  setQaToken(qaToken)
-  if (qaToken.length > 0) sendFrontendMessage('qaModeActive')
-})
-
-ipcMain.on('openAuthModalIfAppReloads', () => {
-  onboardLocalStore.set('openAuthModalIfAppReloads', true)
-})
-
-ipcMain.on('killOverlay', async () => {
-  const hpOverlay = await getHpOverlay()
-  hpOverlay?.closeOverlay()
-})
-
-ipcMain.on('toggleOverlay', async (ev, ...args) => {
-  const hpOverlay = await getHpOverlay()
-  hpOverlay?.toggleOverlay(...args)
-})
-
-ipcMain.handle('getHyperPlayListings', async () => {
-  const listingsMap = await getHyperPlayReleaseObject()
-  return JSON.parse(JSON.stringify(listingsMap))
-})
 /*
  * INSERT OTHER IPC HANDLERS HERE
  */
